修改平均时间的数据类型

+++ blade-service-api/mk-composition-config-api/src/main/java/org/springblade/composition/vo/AutoInspectionVO.java
@@ -10,5 +10,5 @@ public class AutoInspectionVO {
 
 	Integer realSetTotal;
 	Integer correctOrErrorCount;
-	float avgTime;
+	Integer avgTime;
 }
